Consolidate classifier/regressor into single enhanced implementation
Delete the redundant base and MoE-loader classifier/regressor modules and
promote the Enhanced implementation to the public TabLDMClassifier /
TabLDMRegressor names, minimizing impact on existing references in
README, tutorials, and tests. Relocate the shared _moe_load_mismatch
helper into sklearn_utils.py and fix a stale batch_size default in docs.
