+++ README.md
@@ -203,7 +203,7 @@ clf = TabLDMClassifier(
     softmax_temperature=0.9,      # Temperature controlling prediction confidence
     average_logits=True,          # Average logits (True) or probabilities (False)
     support_many_classes=True,    # Automatically handle more than 10 classes
-    batch_size=8,                 # Ensemble members processed together; lower saves memory
+    batch_size=4,                 # Ensemble members processed together; lower saves memory
     kv_cache=False,               # Cache training-data KV projections for repeated inference
     model_path=None,              # Checkpoint path; None downloads from Hugging Face
     allow_auto_download=True,     # Download automatically when not found locally
