+++ README_CN.md
@@ -190,7 +190,7 @@ clf = TabLDMClassifier(
     softmax_temperature=0.9,      # 控制预测置信度的温度参数
     average_logits=True,          # 对 logits（True）或概率（False）取平均
     support_many_classes=True,    # 自动处理类别数超过 10 的任务
-    batch_size=8,                 # 一次处理的集成成员数；调低可节省内存
+    batch_size=4,                 # 一次处理的集成成员数；调低可节省内存
     kv_cache=False,               # 缓存训练数据的 KV 投影，以加速重复推理
     model_path=None,              # 检查点路径；为 None 时从 Hugging Face 下载
     allow_auto_download=True,     # 本地未找到时自动下载
